refactor: consolidate types
