+++ src/api/data/tree/importCommon.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { Connection, SfError, Messages } from '@salesforce/core';
-import { SObjectTreeInput, SObjectTreeFileContents } from '../../../dataSoqlQueryTypes.js';
+import { SObjectTreeInput, SObjectTreeFileContents } from '../../../types.js';
 import { ResponseRefs, TreeResponse } from './importTypes.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
