+++ src/bulkOperationBase.ts
